add search by filter

+++ Titova_laboratornaya1/CPipe.h
@@ -8,7 +8,7 @@ class CPipe
 public:
     static int MaxID;
     std::string name;
-    int length;
+    double length;
     int diametr;
     bool repair;
